MODULES-10362: Handle authenticationinfo consistently.
