+++ lib/puppet/type/iis_site.rb
@@ -2,6 +2,7 @@ require 'puppet/parameter/boolean'
 require_relative '../../puppet_x/puppetlabs/iis/property/name'
 require_relative '../../puppet_x/puppetlabs/iis/property/hash'
 require_relative '../../puppet_x/puppetlabs/iis/property/path'
+require_relative '../../puppet_x/puppetlabs/iis/property/authenticationinfo'
 
 Puppet::Type.newtype(:iis_site) do
   @doc = "Allows creation of a new IIS Web Site and configuration of site
@@ -346,20 +347,7 @@ Puppet::Type.newtype(:iis_site) do
     end
   end
 
-  newproperty(:authenticationinfo) do
-    desc 'Enable and disable authentication schemas. Note: some schemas require
-          additional Windows features to be installed, for example windows
-          authentication. This type does not ensure a given feature is installed
-          before attempting to configure it.'
-    valid_schemas = ['anonymous', 'basic', 'clientCertificateMapping',
-                     'digest', 'iisClientCertificateMapping', 'windows']
-    validate do |value|
-      raise "#{name} should be a Hash" unless value.is_a? ::Hash
-      unless (value.keys & valid_schemas) == value.keys
-        raise('All schemas must specify any of the following: anonymous, basic, clientCertificateMapping, digest, iisClientCertificateMapping, or windows')
-      end
-    end
-  end
+  newproperty(:authenticationinfo, parent: PuppetX::PuppetLabs::IIS::Property::AuthenticationInfo)
 
   autorequire(:iis_application_pool) { self[:applicationpool] }
